voice-call-agent/requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
twilio==9.3.1
python-dotenv==1.0.1
requests==2.32.3
httpx==0.27.2
openai==1.52.2
python-multipart==0.0.20
pypdf==5.1.0
python-docx==1.1.2
psycopg2-binary==2.9.9
livekit-agents>=1.2.0
